requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & ORM
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1

# Background Jobs
rq==1.16.1
redis==5.0.1
taskiq==0.12.1
taskiq-redis==0.5.1

# Event Streaming
confluent-kafka==2.3.0

# Agent Framework
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.10

# Configuration & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# External Integrations
twilio==8.11.1
boto3==1.34.19  # For AWS SES
httpx==0.26.0  # For HTTP clients
suprsend-py-sdk>=0.16.0  # For email/WhatsApp via SuprSend

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytz==2024.1

# Monitoring & Metrics
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # Already listed but needed for TestClient
locust==2.20.0  # Load testing

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
